fix: 버그 수정 (#146) - DB에서 DailyPopularWordLog 타입으로 가져올 때의 타입 불일치 문제해결 - TimeUilts 코드 수정

+++ domain-service/src/main/java/com/domainservice/domain/search/service/dto/vo/DailyPopularWordLog.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public record DailyPopularWordLog(
 	String word,
-	Integer searchedCount,
+	Long searchedCount,
 	LocalDateTime lastSearchedAt
 ) {
 }
